requirements.txt
et_xmlfile==2.0.0
numpy==2.2.4
openpyxl==3.1.5
pandas==2.2.3
py4j==0.10.9.7
pyodbc==5.2.0
pyspark==3.5.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
six==1.17.0
tzdata==2025.2
